+++ src/ModelTypes.h
@@ -52,7 +52,7 @@ struct seed {
  */
 enum class ModelStructureId
 {
-    ORIGINAL,
+    ORIGINAL = 1, // Allows a zero 'Not specified' value
     IRISH
 };
 
@@ -61,7 +61,7 @@ enum class ModelStructureId
  */
 enum class ModelModeId
 {
-	INFERENCE,
+	INFERENCE = 1, // Allows a zero 'Not specified' value
 	PREDICTION
 };
 
@@ -73,7 +73,7 @@ struct ModelInputParameters
 	int herd_id;
 	double tau;
 	int num_threads;
-    ModelStructureId model_structure;
+    ModelStructureId model_structure = ModelStructureId(0);
 	int nsteps;
 	int nParticalLimit;
 	int nSim;
@@ -103,7 +103,7 @@ struct ModelInputParameters
 	double prior_ps_shape2;
 	std::vector<double> posterior_param_list;
 	int posterior_parameter_select;
-	ModelModeId run_type;
+	ModelModeId run_type = ModelModeId(0);
 	std::vector<double> prior_param_list;
 };
