+++ src/Utilities.cpp
@@ -7,7 +7,7 @@ namespace Utilities
 {
     std::string toUpper(std::string str)
     {
-        std::string _temp(str);
+        std::string _temp = str;
         std::transform(_temp.begin(), _temp.end(), _temp.begin(), toupper);
 
         return _temp;
